requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Deep Learning (Optional)
tensorflow==2.13.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# MLOps and Experiment Tracking
mlflow==2.5.0
optuna==3.2.0

# API Development
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3
python-multipart==0.0.6

# Azure ML SDK
azure-ai-ml==1.8.0
azure-identity==1.13.0
azureml-mlflow==1.51.0

# Data Validation
pydantic==2.0.3
great-expectations==0.17.5

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1
hydra-core==1.3.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0
pylint==2.17.4

# Utilities
tqdm==4.65.0
joblib==1.3.1
python-dateutil==2.8.2

# Logging and Monitoring
loguru==0.7.0

# HTTP requests
requests==2.31.0
